Restructure Game class structure
